+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Fixed
 
 - Add error handling for AST parsing in next_parseLayoutJs.ts
+- Fixed scope issue in parseLayoutJs.ts
 
 ## 1.2.2-alpha.7 - 2026.01.24